--- a/source/common/nav.js
+++ b/source/common/nav.js
@@ -27,17 +27,17 @@ const links = [
   {
     label: "Love Compatibility",
     url: "../love-compatibility/",
-    imageUrl: "../home-page/images/love-compatibility-img.png",
+    imageUrl: "../home-page/images/love-compatibility-img.jpg",
   },
   {
     label: "Fortune Cookie",
     url: "../fortune-cookie/",
-    imageUrl: "../home-page/images/fortune-cookie-img.png",
+    imageUrl: "../home-page/images/fortune-cookie-img.jpg",
   },
   {
     label: "Palm Reading",
     url: "../palm-reading/",
-    imageUrl: "../home-page/images/palm-reading-img.png",
+    imageUrl: "../home-page/images/palm-reading-img.jpg",
   },
 ];