--- a/source/fortune-cookie/style.css
+++ b/source/fortune-cookie/style.css
@@ -14,7 +14,7 @@ body {
   right: -10px;
   bottom: -10px;
   z-index: -1;
-  background-image: url("../common/Images/Background.png");
+  background-image: url("../common/Images/Background.jpg");
   background-size: cover;
   background-position: center;
 }