--- a/source/home-page/index.html
+++ b/source/home-page/index.html
@@ -42,10 +42,16 @@
 
     <div class="cards">
       <a href="../love-compatibility/" class="card-container">
-        <img
-          src="images/love-compatibility-img.png"
-          alt="Love Compatibility Card"
-        />
+        <picture>
+          <source
+            srcset="images/love-compatibility-img.webp"
+            type="image/webp"
+          />
+          <img
+            src="images/love-compatibility-img.jpg"
+            alt="Love Compatibility Card"
+          />
+        </picture>
         <span class="description">
           <strong>Love Compatibility</strong>
           <span>
@@ -55,7 +61,10 @@
         </span>
       </a>
       <a href="../fortune-cookie/" class="card-container">
-        <img src="images/fortune-cookie-img.png" alt="Fortune Cookie Card" />
+        <picture>
+          <source srcset="images/fortune-cookie-img.webp" type="image/webp" />
+          <img src="images/fortune-cookie-img.jpg" alt="Fortune Cookie Card" />
+        </picture>
         <span class="description">
           <strong>Fortune Cookie</strong>
           <span>
@@ -65,7 +74,10 @@
         </span>
       </a>
       <a href="../palm-reading/" class="card-container">
-        <img src="images/palm-reading-img.png" alt="Palm Reading Card" />
+        <picture>
+          <source srcset="images/palm-reading-img.webp" type="image/webp" />
+          <img src="images/palm-reading-img.jpg" alt="Palm Reading Card" />
+        </picture>
         <span class="description">
           <strong>Palm Reading</strong>
           <span>