--- a/source/home-page/index.html
+++ b/source/home-page/index.html
@@ -44,6 +44,10 @@
       <a href="../love-compatibility/" class="card-container">
         <picture>
           <source
+            srcset="images/love-compatibility-img.avif"
+            type="image/avif"
+          />
+          <source
             srcset="images/love-compatibility-img.webp"
             type="image/webp"
           />
@@ -62,6 +66,7 @@
       </a>
       <a href="../fortune-cookie/" class="card-container">
         <picture>
+          <source srcset="images/fortune-cookie-img.avif" type="image/avif" />
           <source srcset="images/fortune-cookie-img.webp" type="image/webp" />
           <img src="images/fortune-cookie-img.jpg" alt="Fortune Cookie Card" />
         </picture>
@@ -75,6 +80,7 @@
       </a>
       <a href="../palm-reading/" class="card-container">
         <picture>
+          <source srcset="images/palm-reading-img.avif" type="image/avif" />
           <source srcset="images/palm-reading-img.webp" type="image/webp" />
           <img src="images/palm-reading-img.jpg" alt="Palm Reading Card" />
         </picture>