--- a/source/home-page/styles.css
+++ b/source/home-page/styles.css
@@ -1,7 +1,7 @@
 /* used to set background image*/
 
 body {
-  background-image: url("../common/Images/Background.png");
+  background-image: url("../common/Images/Background.jpg");
   display: flex;
   flex-direction: column;
 }