--- a/source/love-compatibility/index.html
+++ b/source/love-compatibility/index.html
@@ -39,13 +39,16 @@
       </header>
       <div class="wheel-field" id="left-wheel-field">
         <div class="wheel" id="left-wheel">
-          <img
-            src="images/zodiac-wheel.png"
-            class="wheel-img"
-            id="left-wheel-img"
-            alt="left zodiac wheel"
-            draggable="false"
-          />
+          <picture>
+            <source srcset="images/zodiac-wheel.webp" type="image/webp" />
+            <img
+              src="images/zodiac-wheel.png"
+              class="wheel-img"
+              id="left-wheel-img"
+              alt="left zodiac wheel"
+              draggable="false"
+            />
+          </picture>
         </div>
       </div>
       <div class="bday-input" id="left-bday-input">
@@ -112,7 +115,10 @@
       <div class="popup" id="help">
         <div id="text">
           <button class="close-button" id="close-button">
-            <img src="./images/close.png" alt="close" />
+            <picture>
+              <source srcset="images/close.webp" type="image/webp" />
+              <img src="images/close.png" alt="close" />
+            </picture>
           </button>
           <h2 id="pairing">User Guide</h2>
           <h3>Step 1: Selecting Zodiac Pairs</h3>