--- a/source/love-compatibility/index.html
+++ b/source/love-compatibility/index.html
@@ -40,6 +40,7 @@
       <div class="wheel-field" id="left-wheel-field">
         <div class="wheel" id="left-wheel">
           <picture>
+            <source srcset="images/zodiac-wheel.avif" type="image/avif" />
             <source srcset="images/zodiac-wheel.webp" type="image/webp" />
             <img
               src="images/zodiac-wheel.png"
@@ -71,13 +72,17 @@
 
       <div class="wheel-field" id="right-wheel-field">
         <div class="wheel" id="right-wheel">
-          <img
-            src="images/zodiac-wheel.png"
-            class="wheel-img"
-            id="right-wheel-img"
-            alt="right zodiac wheel"
-            draggable="false"
-          />
+          <picture>
+            <source srcset="images/zodiac-wheel.avif" type="image/avif" />
+            <source srcset="images/zodiac-wheel.webp" type="image/webp" />
+            <img
+              src="images/zodiac-wheel.png"
+              class="wheel-img"
+              id="right-wheel-img"
+              alt="right zodiac wheel"
+              draggable="false"
+            />
+          </picture>
         </div>
       </div>
       <div id="arrows">
@@ -116,6 +121,7 @@
         <div id="text">
           <button class="close-button" id="close-button">
             <picture>
+              <source srcset="images/close.avif" type="image/avif" />
               <source srcset="images/close.webp" type="image/webp" />
               <img src="images/close.png" alt="close" />
             </picture>