--- a/source/love-compatibility/styles.css
+++ b/source/love-compatibility/styles.css
@@ -8,7 +8,7 @@ body {
   display: flex;
   align-items: center;
   justify-content: center;
-  background-image: url("../common/Images/Background.png");
+  background-image: url("../common/Images/Background.jpg");
   overflow: visible;
   overflow-x: hidden;
 }