--- a/source/palm-reading/index.html
+++ b/source/palm-reading/index.html
@@ -66,7 +66,15 @@
     <div id="results-frame-anchor">
       <div id="results-frame">
         <h2 id="results-title">Results</h2>
-        <img id="palm-guide" class="align-div" src="images/palm-reading.PNG" />
+        <picture>
+          <source srcset="images/palm-reading.avif" type="image/avif" />
+          <source srcset="images/webp.avif" type="image/webp" />
+          <img
+            id="palm-guide"
+            class="align-div"
+            src="images/palm-reading.PNG"
+          />
+        </picture>
         <div class="tabs-wrapper">
           <div class="tabs">
             <label class="tab" style="--color: #f7ac80">