… source/common/Images/Background.avif
… source/common/Images/Background.jpg
… source/common/Images/Background.webp
… source/fortune-cookie/images/transparent-cookie-left.avif
… source/fortune-cookie/images/transparent-cookie-left.png
… source/fortune-cookie/images/transparent-cookie-left.webp
… source/fortune-cookie/images/transparent-cookie-right.avif
… source/fortune-cookie/images/transparent-cookie-right.png
… source/fortune-cookie/images/transparent-cookie-right.webp
… source/fortune-cookie/images/transparent-cookie.avif
… source/fortune-cookie/images/transparent-cookie.png
… source/fortune-cookie/images/transparent-cookie.webp
… source/home-page/images/fortune-cookie-img.avif
… source/home-page/images/fortune-cookie-img.jpg
… source/home-page/images/fortune-cookie-img.png
… source/home-page/images/fortune-cookie-img.webp
… source/home-page/images/love-compatibility-img.avif
… source/home-page/images/love-compatibility-img.jpg
… source/home-page/images/love-compatibility-img.png
… source/home-page/images/love-compatibility-img.webp
… source/home-page/images/palm-reading-img.avif
… source/home-page/images/palm-reading-img.jpg
… source/home-page/images/palm-reading-img.webp
… source/love-compatibility/images/arrow-left.avif
… source/love-compatibility/images/arrow-right.avif
… source/love-compatibility/images/close.avif
… source/love-compatibility/images/close.png
… source/love-compatibility/images/close.webp
… source/love-compatibility/images/zodiac-wheel.avif
… source/love-compatibility/images/zodiac-wheel.png