There's still more performance work to do, but I'd like to merge the current progress into `main`.

Improved the loading time of the site on `Good 3G (DL: 1536 Kbps, UP: 750 Kbps, latency 40ms)` from 35.79 seconds to 4.27 seconds by simply replacing usage of pngs with jpegs (which should be just as compatible) and also adding the `<picture>` tag to use webp when possible.

See my [commit message](https://github.com/20-20REENVISIONED/Refactored-fortune-teller/commit/99394a49389b737024291141c53f4141ea8e72c8) for more details.